折り紙を登録するカスタムフック
